clear out unused components, consolidate lib files
