--- src/feat/Projects/index.ts
@@ -1,5 +1,0 @@
-export * from './ProjectViewer';
-export * from './ProjectProvider';
-export * from './hooks/useProjectViewer';
-export * from './ui/ProjectItem';
-export * from './types';
