+++ src/index.html
@@ -1,13 +1,10 @@
 <!doctype html>
 <html>
-
-<head>
-  <meta charset="UTF-8" />
-  <title><%- title %></title>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-
+  <head>
+    <meta charset="UTF-8" />
+    <title><%- title %></title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
