--- src/lib/canvas/index.ts
@@ -1,2 +1,0 @@
-export * from './Canvas';
-export * from './types';
