--- src/lib/errors/helpers.ts
@@ -1,4 +1,0 @@
-import { ConvexError } from 'convex/values';
-
-export const isNotFoundError = (error: unknown) =>
-  error instanceof ConvexError && error.data.code === 404;
