--- src/lib/errors/index.ts
@@ -1,3 +1,0 @@
-export * from './ErrorBoundary';
-export * from './Fallback';
-export * from './helpers';
